refactor: extrair caso selectNewGame p/ component
