+++ package.json
@@ -13,7 +13,8 @@
         "date-fns": "^2.29.3",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "react-router-dom": "^6.8.1"
+        "react-router-dom": "^6.8.1",
+        "ts-pattern": "^4.1.4"
     },
     "devDependencies": {
         "@types/react": "^18.0.27",
