+++ src/components/GameTable/components/index.ts
@@ -1,2 +1,3 @@
 export * from './EndGame';
 export * from './HiddenGame';
+export * from './SelectNewGameMessage';
